pyproject.toml
[project]
name = "spotify-exploration"
version = "0.1.0"
description = "A personal Spotify Wrapped dashboard built with Streamlit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.32.0",
    "spotipy>=2.23.0",
    "pandas>=2.0.0",
    "plotly>=5.18.0",
    "python-dotenv>=1.0.0",
]
